frontend/ml/requirements.txt
# AgriSense AI – Python ML Dependencies
# =====================================
# Install with: pip install -r requirements.txt

# Deep Learning – Disease Detection
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0

# OpenAI Integration
openai>=1.0.0

# Machine Learning – Yield Prediction
scikit-learn>=1.3.0
xgboost>=2.0.0
numpy>=1.26.0
pandas>=2.0.0
joblib>=1.3.0

# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
